+++ src/routes/blog.svelte
@@ -3,6 +3,32 @@
    import posts from './blog/posts.json';
 
    let count = posts.length;
+   import supabase from '$lib/db';
+   import { onMount } from "svelte";
+
+   async function addUniqueVisitor(path){
+      let request = await fetch('https://api.ipify.org?format=json')
+      let json = await request.json()
+      let request_geo = await fetch('http://ip-api.com/json/' + json.ip)
+      let json_geo = await request_geo.json()
+      let location = `${json_geo.country}, ${json_geo.regionName}, ${json_geo.city}`;
+      const { data, error } = await supabase.from('unique_visitors').insert([{
+         ip_address:json.ip,
+         url_position:path,
+         location: location
+      }])
+   }
+   onMount(()=>{
+      console.log('mounted')
+      let cookie = window.localStorage.getItem('unique');
+      console.log(cookie)
+      let path = window.location.pathname
+      if(!cookie){
+         console.log('if cookie')
+         addUniqueVisitor(path)
+         window.localStorage.setItem('unique', 'visited')
+      }
+   })
 </script>
 
 <svelte:head>
